refactor: delete orphan re-export barrels and single-caller wrappers
Removes five small structural-overlap findings from a debt audit of
module/responsibility boundaries: two 1-2-consumer re-export shims
(src/tools/userQuestion/index.ts, src/auth/codex/codexPkce.ts) whose
callers now import the defining module directly; the unused
CODEX_*_TOOL re-export block in src/tools/codexShared.ts, whose two
test-kernel consumers now import the canonical @shared/schemas/codex
path already used by the extension frontend; desktopSettingsIpcHelpers.ts,
four single-caller lazy-import wrappers banned by CLAUDE.md's
"single-caller extractions are banned" rule, inlined at their one call
site in packages/desktop/src/main/index.ts while preserving the
dynamic-import bundle-splitting behavior; and a duplicated REPO_ROOT
path computation in two test-kernel files, now sourced from the single
canonical export in src/test-kernel/support/repoScan.ts.

No behavior change — every replacement targets a symbol that already
canonically exists, verified by exhaustive grep for all consumers of
each deleted binding.
