--- packages/desktop/src/main/desktopSettingsIpcHelpers.ts
@@ -1,43 +1,0 @@
-import type { ToolTerminalAction } from '@controllers/settingsView/ToolDashboardData';
-import type {
-  ToolCommandKind,
-  ToolDashboardItem,
-} from '@shared/schemas/settingsViewMessages';
-import type { ExternalToolCheckResult } from '@tools/toolAvailability';
-
-/**
- * Module-level helpers for the desktop settings IPC. These have no closure
- * dependencies on the IPC instance, so they live here to keep the factory in
- * `desktopSettingsIpc.ts` focused on request wiring. Each tool-availability
- * helper defers its tool-layer import so the heavy module only loads when a
- * tool dashboard request actually arrives.
- */
-
-export async function buildDefaultToolDashboardItems(
-  cachedResults?: ExternalToolCheckResult[],
-): Promise<ToolDashboardItem[]> {
-  const { buildToolDashboardItems } =
-    await import('@controllers/settingsView/ToolDashboardData');
-  return buildToolDashboardItems(cachedResults);
-}
-
-export async function getCachedToolCheckResults(): Promise<
-  ExternalToolCheckResult[] | undefined
-> {
-  const { getLastCheckResults } = await import('@tools/toolAvailability');
-  return getLastCheckResults() ?? undefined;
-}
-
-export async function refreshDefaultToolAvailability(): Promise<void> {
-  const { refreshToolAvailability } = await import('@tools/toolAvailability');
-  await refreshToolAvailability();
-}
-
-export async function planDefaultToolTerminalAction(
-  toolId: string,
-  kind: ToolCommandKind,
-): Promise<ToolTerminalAction> {
-  const { planToolTerminalAction } =
-    await import('@controllers/settingsView/ToolDashboardData');
-  return planToolTerminalAction({ toolId, commandKind: kind });
-}
