--- src/auth/codex/codexPkce.ts
@@ -1,10 +1,0 @@
-/**
- * Re-export shared PKCE helpers. Kept as `@auth/codex` path for existing
- * imports; implementation lives in `@auth/oauth`.
- */
-export {
-  generateCodeVerifier,
-  computeCodeChallenge,
-  generatePkcePair,
-  generateOAuthState,
-} from '../oauth/pkce';
