+++ src/auth/codex/index.ts
@@ -19,12 +19,6 @@ export {
   CODEX_ORIGINATOR_HEADER,
   CODEX_SESSION_SECRET_KEY,
 } from './codexConstants';
-export {
-  generateCodeVerifier,
-  computeCodeChallenge,
-  generatePkcePair,
-  generateOAuthState,
-} from './codexPkce';
 export { extractCodexClaims } from './codexJwt';
 export {
   CodexAuthError,
