--- src/tools/userQuestion/index.ts
@@ -1,1 +1,0 @@
-export { AskUserQuestionTool } from './UserQuestionTool';
